refactor: simplify generic tracker class f97ad57
> rename appropriately to `Counter` as this reflects functionality
> remove unnecessary `MultiValue` class. Duplicate as simple `Counter` performs same functionality
> add separate counter class that maintains previous counter histories
