+++ lib/src/core/yaml_transformers/counter/counter_with_history.dart
@@ -1,57 +1,5 @@
 part of 'transform_counter.dart';
 
-base class CounterWithNoHistory<T> {
-  CounterWithNoHistory() : _counter = {};
-
-  /// A tracker keeping count of each value being added to it
-  final Map<TrackerKey<T>, int> _counter;
-
-  /// Get the current state of the tracker
-  Map<TrackerKey<T>, int> get trackerState => _counter;
-
-  /// Creates a tracker key tracking
-  TrackerKey<T> createKey(T value, {Origin? origin}) {
-    return TrackerKey<T>.fromValue(value, origin!);
-  }
-
-  /// Adds a key if missing and increments the count if present.
-  void _addKey(TrackerKey<T> key, {bool isStartingTracker = false}) {
-    _counter.update(
-      key,
-      (value) => ++value,
-      ifAbsent: () => isStartingTracker ? 0 : 1,
-    );
-  }
-
-  /// Set up all trackable keys.
-  ///
-  /// Important when we want to track an exact number for each key/value
-  void prefill(List<dynamic> keys, {Origin? origin}) {
-    for (final value in keys) {
-      final key = createKey(value as T, origin: origin);
-      _addKey(key, isStartingTracker: true);
-    }
-  }
-
-  /// Increment with dynamic value
-  void increment(Iterable<dynamic> values, {Origin? origin}) {
-    for (final candidate in values) {
-      final key = createKey(candidate as T, origin: origin);
-      _addKey(key);
-    }
-  }
-
-  /// 
-  int getCount(T value, {required Origin origin}) {
-    final key = createKey(value); // Key to get count using
-    return getCountFromKey(key);
-  }
-
-  int getCountFromKey(TrackerKey<T> key) {
-    return _counter[key] ?? 0; // Return 0 if missing
-  }
-}
-
 /// A generic tracker class that tracks a specific values and also maintain the
 /// history of previously tracked value based on a cursor.
 ///
@@ -63,8 +11,8 @@ base class CounterWithNoHistory<T> {
 /// Generic [T] - denotes strict type of value being tracked.
 ///
 /// See [TrackerKey]
-base class SingleValueCounter<C, T> extends CounterWithNoHistory<T> {
-  SingleValueCounter() : _counterHistory = {};
+base class CounterWithHistory<C, T> extends Counter<T> {
+  CounterWithHistory() : _counterHistory = {};
 
   /// History of counters.
   ///
