--- lib/src/core/yaml_transformers/tranform_counter/transform_counter.dart
@@ -1,8 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:magical_version_bump/src/utils/enums/enums.dart';
-import 'package:magical_version_bump/src/utils/exceptions/command_exceptions.dart';
-import 'package:meta/meta.dart';
-
-part 'multi_value_counter.dart';
-part 'single_value_counter.dart';
-part 'tracker_key.dart';
